+++ apps/common-app/src/apps/reanimated/examples/WebWorkerExample.tsx
@@ -1,9 +1,55 @@
 import React, { useEffect, useState } from 'react';
 import { Button, StyleSheet, Text, View } from 'react-native';
-import { myWorker, performHeavyComputation } from './WebWorker';
+import { WebWorker } from 'react-native-worklets';
+
+export function performHeavyComputation() {
+  console.log('Starting heavy computation');
+  for (let i = 0; i < 500_000_000; i++) {
+    Math.sqrt(i);
+  }
+  console.log('Done');
+}
 
 export default function WebWorkerExample() {
   const [counter, setCounter] = useState(0);
+  const [myWorker, setMyWorker] = useState<WebWorker<string, Record<string, unknown>> | null>(null);
+  
+  useEffect(() => { 
+    const myWorker = new WebWorker<string, Record<string, unknown>>(
+      'Worker #1',
+    () => {
+      'worklet';
+      let state = 0;
+
+      onmessage = ({ data }: { data: string }) => {
+        state = state + 1;
+        console.log(`Worker started job ${state} with data: ${data}`);
+        if (data === 'error') {
+          throw new Error('Error from worker!!!');
+        }
+        console.log('Starting heavy computation');
+        for (let i = 0; i < 500_000_000; i++) {
+          Math.sqrt(i);
+        }
+        global.postMessage(`Done job ${state}`);
+      };
+      },
+    );
+
+    myWorker.onmessage = (e: { data: Record<string, unknown> }) => {
+      console.log('JS received:', e.data);
+    };
+
+    myWorker.onerror = (e: { data: string }) => {
+      console.error('JS received error:', e);
+    };
+
+    setMyWorker(myWorker);
+
+    return () => {
+      myWorker.terminate();
+    };
+  }, []);
 
   useEffect(() => {
     const interval = setInterval(() => {
@@ -13,16 +59,15 @@ export default function WebWorkerExample() {
     return () => clearInterval(interval);
   }, []);
 
-
   return (
     <View style={styles.marginTop}>
-      <View style={styles.counterContainer}>  
+      <View style={styles.counterContainer}>
         <Text>Counter on main thread: {counter}</Text>
       </View>
 
       <Button
         onPress={() => {
-          myWorker.postMessage(counter.toString());
+          myWorker?.postMessage(counter.toString());
         }}
         title="Send message to worker (run heavy computation on worker)"
       />
@@ -36,14 +81,14 @@ export default function WebWorkerExample() {
 
       <Button
         onPress={() => {
-          myWorker.terminate();
+          myWorker?.terminate();
         }}
         title="Terminate worker (finish current job and terminate)"
       />
 
       <Button
         onPress={() => {
-          myWorker.postMessage('error');
+          myWorker?.postMessage('error');
         }}
         title="Trigger error in worker"
       />
